:boom: move files from image_app to uid
