+++ django-data/image/uid/fixtures/image_app/animal.json
