+++ django-data/image/uid/fixtures/image_app/dictcountry.json
