+++ django-data/image/uid/fixtures/image_app/dictstage.json
