+++ django-data/image/uid/fixtures/image_app/dictuberon.json
