+++ django-data/image/uid/fixtures/image_app/name.json
