+++ django-data/image/uid/fixtures/image_app/ontology.json
