+++ django-data/image/uid/fixtures/image_app/publication.json
