+++ django-data/image/uid/fixtures/image_app/submission.json
