+++ django-data/image/uid/management/commands/clean_backup.py
