+++ django-data/image/uid/management/commands/initializedb.py
