+++ django-data/image/uid/management/commands/truncate_image_tables.py
