+++ django-data/image/uid/templates/image_app/dashboard.html
