+++ django-data/image/uid/templates/image_app/uploading_data.html
