+++ django-data/image/uid/tests/test_commands.py
